requirements.txt
pandas
python-dotenv
pydantic
fastapi
streamlit
googlemaps
populartimes
sqlalchemy